refactor: consolidate 5 OpenAI-compat engine wrappers into single data-driven file
Replace vllm.py, sglang.py, llamacpp.py, mlx.py, lmstudio.py with
openai_compat_engines.py using dynamic class creation and registration.
