+++ tests/engine/test_engine_model_matrix.py
@@ -9,7 +9,7 @@ from openjarvis.core.registry import EngineRegistry
 from openjarvis.core.types import Message, Role
 from openjarvis.engine._base import EngineConnectionError
 from openjarvis.engine.ollama import OllamaEngine
-from openjarvis.engine.vllm import VLLMEngine
+from openjarvis.engine.openai_compat_engines import VLLMEngine
 
 ENGINES_AND_HOSTS = [
     ("vllm", "http://testhost:8000"),
